# Push command performance improvements