--- a/lean/components/cloud/push_manager.py
+++ b/lean/components/cloud/push_manager.py
@@ -43,6 +43,8 @@ class PushManager:
         self._project_manager = project_manager
         self._project_config_manager = project_config_manager
         self._last_file = None
+        self._cloud_projects = []
+        self._lean_environments = None
 
     def push_projects(self, projects_to_push: List[Path], organization_id: Optional[str] = None) -> None:
         """Pushes the given projects from the local drive to the cloud.
@@ -56,17 +58,13 @@ class PushManager:
                                        for project in projects_to_push
                                        for library in self._project_manager.get_project_libraries(project)]
         projects = sorted(projects)
-
-        cloud_projects = self._api_client.projects.get_all()
-        environments = self._api_client.lean.environments()
-
         pushed_projects = {}
 
         for index, project in enumerate(projects, start=1):
             relative_path = project.relative_to(Path.cwd())
             try:
                 self._logger.info(f"[{index}/{len(projects)}] Pushing '{relative_path}'")
-                pushed_project = self._push_project(project, cloud_projects, organization_id, environments)
+                pushed_project = self._push_project(project, organization_id)
                 pushed_projects[project] = pushed_project
             except Exception as ex:
                 self._logger.debug(traceback.format_exc().strip())
@@ -75,19 +73,14 @@ class PushManager:
                 else:
                     self._logger.warn(f"Cannot push '{relative_path}': {ex}")
 
-        pushed_cloud_projects = pushed_projects.values()
-        cloud_projects = [project for project in cloud_projects if project.projectId not in pushed_cloud_projects]
-        cloud_projects.extend(pushed_cloud_projects)
+        self._update_cloud_library_references(pushed_projects)
 
-        self._update_cloud_library_references(pushed_projects, cloud_projects)
-
-    def _update_cloud_library_references(self, projects: Dict[Path, QCProject],
-                                         cloud_projects: List[QCProject]) -> None:
+    def _update_cloud_library_references(self, projects: Dict[Path, QCProject]) -> None:
         for path, project in projects.items():
             local_libraries_cloud_ids = self._get_local_libraries_cloud_ids(path)
 
-            self._add_new_libraries(project, local_libraries_cloud_ids, cloud_projects)
-            self._remove_outdated_libraries(project, local_libraries_cloud_ids, cloud_projects)
+            self._add_new_libraries(project, local_libraries_cloud_ids)
+            self._remove_outdated_libraries(project, local_libraries_cloud_ids)
 
     def _get_local_libraries_cloud_ids(self, project_dir: Path) -> List[int]:
         project_config = self._project_config_manager.get_project_config(project_dir)
@@ -100,14 +93,10 @@ class PushManager:
 
         return local_libraries_cloud_ids
 
-    @staticmethod
-    def _get_library_name(library_cloud_id: int, cloud_projects: List[QCProject]) -> str:
-        return [project.name for project in cloud_projects if project.projectId == library_cloud_id][0]
+    def _get_library_name(self, library_cloud_id: int) -> str:
+        return self._get_cloud_project(library_cloud_id).name
 
-    def _add_new_libraries(self,
-                           project: QCProject,
-                           local_libraries_cloud_ids: List[int],
-                           cloud_projects: List[QCProject]) -> None:
+    def _add_new_libraries(self, project: QCProject, local_libraries_cloud_ids: List[int]) -> None:
         libraries_to_add = [library_id for library_id in local_libraries_cloud_ids if
                             library_id not in project.libraries]
 
@@ -115,15 +104,12 @@ class PushManager:
             self._logger.info(f"Adding libraries to project {project.name} in the cloud")
 
         for i, library_cloud_id in enumerate(libraries_to_add, start=1):
-            library_name = self._get_library_name(library_cloud_id, cloud_projects)
+            library_name = self._get_library_name(library_cloud_id)
             self._logger.info(f"[{i}/{len(libraries_to_add)}] "
                               f"Adding library {library_name} to project {project.name} in the cloud")
             self._api_client.projects.add_library(project.projectId, library_cloud_id)
 
-    def _remove_outdated_libraries(self,
-                                   project: QCProject,
-                                   local_libraries_cloud_ids: List[int],
-                                   cloud_projects: List[QCProject]) -> None:
+    def _remove_outdated_libraries(self, project: QCProject, local_libraries_cloud_ids: List[int]) -> None:
         libraries_to_remove = [library_id for library_id in project.libraries
                                if library_id not in local_libraries_cloud_ids]
 
@@ -131,36 +117,28 @@ class PushManager:
             self._logger.info(f"Removing libraries from project {project.name} in the cloud")
 
         for i, library_cloud_id in enumerate(libraries_to_remove, start=1):
-            library_name = self._get_library_name(library_cloud_id, cloud_projects)
+            library_name = self._get_library_name(library_cloud_id)
             self._logger.info(f"[{i}/{len(libraries_to_remove)}] "
                               f"Removing library {library_name} from project {project.name} in the cloud")
             self._api_client.projects.delete_library(project.projectId, library_cloud_id)
 
-    def _push_project(self,
-                      project: Path,
-                      cloud_projects: List[QCProject],
-                      organization_id: Optional[str],
-                      environments: List[QCLeanEnvironment]) -> QCProject:
+    def _push_project(self, project: Path, organization_id: Optional[str]) -> QCProject:
         """Pushes a single local project to the cloud.
 
         Raises an error with a descriptive message if the project cannot be pushed.
 
         :param project: the local project to push
-        :param cloud_projects: a list containing all of the user's cloud projects
         :param organization_id: the id of the organization to push the project to
-        :param environments: list of available lean environments
         """
         project_name = project.relative_to(Path.cwd()).as_posix()
 
         project_config = self._project_config_manager.get_project_config(project)
         cloud_id = project_config.get("cloud-id")
 
-        cloud_project_by_id = next(iter([p for p in cloud_projects if p.projectId == cloud_id]), None)
-
         # Find the cloud project to push the files to
-        if cloud_project_by_id is not None:
+        if cloud_id is not None:
             # Project has cloud id which matches cloud project, update cloud project
-            cloud_project = cloud_project_by_id
+            cloud_project = self._get_cloud_project(cloud_id)
         else:
             # Project has invalid cloud id or no cloud id at all, create new cloud project
             new_project = self._api_client.projects.create(project_name,
@@ -174,6 +152,7 @@ class PushManager:
 
             # We need to retrieve the created project again to get all project details
             cloud_project = self._api_client.projects.get(new_project.projectId)
+            self._cloud_projects.append(cloud_project)
 
             # set organization-id in project config
             project_config.set("organization-id", cloud_project.organizationId)
@@ -182,7 +161,7 @@ class PushManager:
         self._push_files(project, cloud_project)
 
         # Finalize pushing by updating locally modified metadata
-        self._push_metadata(project, cloud_project, environments)
+        self._push_metadata(project, cloud_project)
 
         return cloud_project
 
@@ -225,7 +204,7 @@ class PushManager:
 
         self._last_file = None
 
-    def _push_metadata(self, project: Path, cloud_project: QCProject, environments: List[QCLeanEnvironment]) -> None:
+    def _push_metadata(self, project: Path, cloud_project: QCProject) -> None:
         """Pushes local project description and parameters to the cloud.
 
         Does nothing if the cloud is already up-to-date.
@@ -245,7 +224,7 @@ class PushManager:
         local_lean_version = int(project_config.get("lean-engine", "-1"))
         cloud_lean_version = cloud_project.leanVersionId
 
-        default_lean_venv = next((env.id for env in environments if env.path is None), None)
+        default_lean_venv = self._default_environment
         local_lean_venv = project_config.get("python-venv", default_lean_venv)
         cloud_lean_venv = cloud_project.leanEnvironment
 
@@ -267,3 +246,18 @@ class PushManager:
         if update_args != {}:
             self._api_client.projects.update(cloud_project.projectId, **update_args)
             self._logger.info(f"Successfully updated {' and '.join(update_args.keys())} for '{cloud_project.name}'")
+
+    def _get_cloud_project(self, project_id: int) -> QCProject:
+        project = next(iter(p for p in self._cloud_projects if p.projectId == project_id), None)
+        if project is None:
+            project = self._api_client.projects.get(project_id)
+            self._cloud_projects.append(project)
+
+        return project
+
+    @property
+    def _default_environment(self) -> List[QCLeanEnvironment]:
+        if self._lean_environments is None:
+            self._lean_environments = self._api_client.lean.environments()
+
+        return next((env.id for env in self._lean_environments if env.path is None), None)