… tests/components/cloud/test_cloud_project_manager.py
… tests/components/util/test_push_manager.py