Push command performance improvements by removing unnecessary API calls like `projects/read` (all) and `environments`.

## Push times before the improvements

### Newly created project

#### Push 1:
```
lean-cli\root> lean cloud push --project .\test_projects\Proyecto04
[1/1] Pushing 'test_projects\Proyecto04'
Successfully created cloud project 'test_projects/Proyecto04'
Successfully updated cloud file 'test_projects/Proyecto04/main.py'
Successfully updated cloud file 'test_projects/Proyecto04/research.ipynb'
Finished pulling projects in 5.683187007904053 seconds
```

#### Push 2:
```
lean-cli\root> lean cloud push --project .\test_projects\Proyecto04
[1/1] Pushing 'test_projects\Proyecto04'
Successfully created cloud project 'test_projects/Proyecto04'
Successfully updated cloud file 'test_projects/Proyecto04/main.py'
Successfully updated cloud file 'test_projects/Proyecto04/research.ipynb'
Finished pulling projects in 5.764963150024414 seconds
```

#### Push 3:
```
lean-cli\root> lean cloud push --project .\test_projects\Proyecto04
[1/1] Pushing 'test_projects\Proyecto04'
Successfully created cloud project 'test_projects/Proyecto04'
Successfully updated cloud file 'test_projects/Proyecto04/main.py'
Successfully updated cloud file 'test_projects/Proyecto04/research.ipynb'
Finished pulling projects in 5.719933032989502 seconds
```

## Push times after the improvements

### Newly created project

#### Push 1:
```
lean-cli\root> lean cloud push --project .\test_projects\Proyecto04
[1/1] Pushing 'test_projects\Proyecto04'
Successfully created cloud project 'test_projects/Proyecto04'
Successfully updated cloud file 'test_projects/Proyecto04/main.py'
Successfully updated cloud file 'test_projects/Proyecto04/research.ipynb'
Finished pulling projects in 3.214686393737793 seconds
```

#### Push 2:
```
lean-cli\root> lean cloud push --project .\test_projects\Proyecto04
[1/1] Pushing 'test_projects\Proyecto04'
Successfully created cloud project 'test_projects/Proyecto04'
Successfully updated cloud file 'test_projects/Proyecto04/main.py'
Successfully updated cloud file 'test_projects/Proyecto04/research.ipynb'
Finished pulling projects in 3.3903634548187256 seconds
```

#### Push 3:
```
lean-cli\root> lean cloud push --project .\test_projects\Proyecto04
[1/1] Pushing 'test_projects\Proyecto04'
Successfully created cloud project 'test_projects/Proyecto04'
Successfully updated cloud file 'test_projects/Proyecto04/main.py'
Successfully updated cloud file 'test_projects/Proyecto04/research.ipynb'
Finished pulling projects in 3.286642551422119 seconds
```